add necessary modules to core

See merge request ia-projekt-zespolowy-2022-2023/graph-editor-core!1

+++ src/main/java/graph_editor/geometry/Point.java
@@ -2,18 +2,18 @@ package graph_editor.geometry;
 
 // Immutable.
 public class Point implements Geometry {
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 }
